+++ liushu-core/src/engine/candidates.rs
@@ -3,5 +3,5 @@ pub struct Candidate {
     pub text: String,
     pub code: String,
     pub comment: Option<String>,
-    pub weight: u64,
+    pub weight: u32,
 }
